Add files via upload
This is the working code for running an I2C 20x4 LCD off the MEGA BMS master shield.
It is Dave's code from April, I've added very basic display outputs. I think it would be best if Dave or Duncan worked on merging the display features from the protoboard version into this code.