Updated for v2 comms
Initial revision to work with v2 style comms with separate requests for voltage and temperature data. Should still work with v1 style as well.
Revised the constructor to take arguments for the number of cells, baudrate, and comms version to use.
In theory v2 comms should work for up to 84 cells in one string.